Exported some functions
