+++ plugin.go
@@ -1,6 +1,8 @@
 package main
 
 import (
+	"github.com/blinkops/blink-sdk/plugin"
+	"net/http"
 	"os"
 	"path"
 
@@ -29,14 +31,44 @@ func main() {
 		panic(err)
 	}
 
-	plugin, err := implementation.NewHTTPPlugin(currentDirectory)
+	supportedActions := map[string]implementation.ActionHandler{
+		"get":    executeHTTPGetAction,
+		"post":   executeHTTPPostAction,
+		"put":    executeHTTPPutAction,
+		"delete": executeHTTPDeleteAction,
+		"patch":  executeHTTPPatchAction,
+	}
+
+	httpPlugin, err := implementation.NewHTTPPlugin(currentDirectory, supportedActions)
 	if err != nil {
 		log.Error("Failed to create plugin implementation: ", err)
 		panic(err)
 	}
 
-	err = blinkSdk.Start(plugin)
+	err = blinkSdk.Start(httpPlugin)
 	if err != nil {
 		log.Fatal("Error during server startup: ", err)
 	}
 }
+
+
+func executeHTTPGetAction(ctx *plugin.ActionContext, request *plugin.ExecuteActionRequest) ([]byte, error) {
+	return implementation.ExecuteCoreHTTPAction(ctx, http.MethodGet, request)
+}
+
+func executeHTTPPostAction(ctx *plugin.ActionContext, request *plugin.ExecuteActionRequest) ([]byte, error) {
+	return implementation.ExecuteCoreHTTPAction(ctx, http.MethodPost, request)
+}
+
+func executeHTTPPutAction(ctx *plugin.ActionContext, request *plugin.ExecuteActionRequest) ([]byte, error) {
+	return implementation.ExecuteCoreHTTPAction(ctx, http.MethodPut, request)
+}
+
+func executeHTTPDeleteAction(ctx *plugin.ActionContext, request *plugin.ExecuteActionRequest) ([]byte, error) {
+	return implementation.ExecuteCoreHTTPAction(ctx, http.MethodDelete, request)
+}
+
+func executeHTTPPatchAction(ctx *plugin.ActionContext, request *plugin.ExecuteActionRequest) ([]byte, error) {
+	return implementation.ExecuteCoreHTTPAction(ctx, http.MethodPatch, request)
+}
+
